+++ libs/game/tilemap.ts
@@ -22,7 +22,7 @@ namespace tiles {
     }
 
     export class TileMap {
-        camera: game.Camera;
+        camera: scenes.Camera;
         tileWidth: number;
         tileHeight: number;
 
@@ -33,7 +33,7 @@ namespace tiles {
         private _tiles: Tile[];
         private _tileSprites: TileSprite[];
 
-        constructor(camera: game.Camera, tileWidth: number, tileHeight: number) {
+        constructor(camera: scenes.Camera, tileWidth: number, tileHeight: number) {
             this.camera = camera;
             this.tileWidth = tileWidth;
             this.tileHeight = tileHeight;
@@ -83,7 +83,7 @@ namespace tiles {
             return this._tiles[index] = new Tile(img, false);
         }
 
-        render(camera: game.Camera) {
+        render(camera: scenes.Camera) {
             if (!this._map) return;
             if (game.debug) {
                 const offsetX = -camera.offsetX;
@@ -110,7 +110,7 @@ namespace tiles {
             }
         }
 
-        public update(camera: game.Camera) {
+        public update(camera: scenes.Camera) {
             if (!this._map || !this.needsUpdate) return;
 
             this.needsUpdate = false;
